IMS: Check for binderstate before calling endSession
When device is configured to L + G configuration and DDS is switched
to SUB1 -> SUB2 and then SUB2 -> SUB1 observed that user not able to
call on GSM subscription.

Above issue is due to ImsRegistrationListenerProxy listener which
is registered as part of ImsManager.open is not cleared even after moving
feature state to NOT_AVAILABLE because of endSession is never called
for feature.

To fix this issue check for binder state instead of feature state READY
before calling end session. Note that end session will be triggered after
moving feature state to NOT_AVAILABLE.

Notifying ims state change always to the clients from imsStatusChanged
due to which end up in calling ImsManager.open from ImsPhoneCallTracker
multiple times as ImsServiceSub supports both MMTEL and EMERGENCY features.
So notify ims state change only for the supported feature state change to
clients.

Change-Id: I39130f68d2af88919bd3d458ea4bacab161cf4e5
CRs-Fixed: 2062477